MCR-3084 Separate test runner class from actual test cases
